sheet-table/requirements.txt
aiogram==3.13.1
gspread==6.1.2
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
